+++ CHANGELOG.md
@@ -1,3 +1,39 @@
+# v14.0.15 (Fri Jan 03 2025)
+
+#### 🐛 Bug Fix
+
+- feat(ONYX-1459): add bottom icon position and dark variant to message component [#297](https://github.com/artsy/palette-mobile/pull/297) ([@dariakoko](https://github.com/dariakoko))
+
+#### Authors: 1
+
+- Daria Kozlova ([@dariakoko](https://github.com/dariakoko))
+
+---
+
+# v14.0.14 (Tue Dec 31 2024)
+
+#### 🐛 Bug Fix
+
+- feat: add a11y to checkbox + testid [#296](https://github.com/artsy/palette-mobile/pull/296) ([@gkartalis](https://github.com/gkartalis))
+
+#### Authors: 1
+
+- George ([@gkartalis](https://github.com/gkartalis))
+
+---
+
+# v14.0.13 (Wed Dec 18 2024)
+
+#### 🐛 Bug Fix
+
+- feat: allow VisualClue color to be customized [#295](https://github.com/artsy/palette-mobile/pull/295) ([@anandaroop](https://github.com/anandaroop))
+
+#### Authors: 1
+
+- Anandaroop Roy ([@anandaroop](https://github.com/anandaroop))
+
+---
+
 # v14.0.12 (Fri Dec 06 2024)
 
 #### 🐛 Bug Fix